requirements.txt
flask
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
pika
python-dotenv
rx
redis
geopy
